requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]==3.11
scikit-learn
certifi
matplotlib
seaborn
xgboost


dill
pickle
scipy
pyyaml


##-e .